+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 - `Stream::curr_char`.
 
+### Fixed
+- Declaration attributes with mixed quotes parsing.
+
 ## [0.8.1] - 2019-01-02
 ### Changed
 - Changed the crate category in the Cargo.toml